Tree-To-HTML Exercise
Tree to HTML
You've been learning about the relationship between data trees and HTML. 

I want you to build this tree with HTML.

How to Complete this Quiz
Take a close look at the relationships in the tree. Pay attention to parents, children and siblings. Notice that some elements have text content.
Download the file called tree-to-html.zip from the Supporting Materials below. Inside you'll find index.html and an image of the tree from above.
Turn on the Udacity Feedback Chrome Extension and edit the index.html file so that the HTML has the same structure as the tree.
When your HTML has the same structure as the tree, the Chrome extension will give you a code. Paste it in the classroom to complete the quiz!