Create index.html
IMAGES! Images on the web are awesome. Time to make one.

For this quiz, you'll be creating an image here in the Udacity classroom. I'll give you the URL of an image and it will be your job to make it display.

An image is made with an  element. It looks like so:

<img src="http://somewebsite.com/image.jpg" alt="short description">
The source attribute, src, is like the href of a link - it is the URL of the image you want to display. For now, your images will need to be hosted online, which means that the URL will need to start with http:// or https://. You'll learn about another way to set image source in the next quiz. The alt attribute stands for "alternative description," which is important for people who use screen readers to browse the web. This is text that will show up in lieu of the actual image.

An image element is a little different than the elements you've seen before. Images do not need closing tags!