Add files via upload
Figures
It's important to respect copyright and attribution on the web. When you use someone else's work, you need to give the author credit. With that in mind, for this quiz you're going to create an image with a caption underneath. I want you to give credit to the image's photographer by attributing them in the image's caption and providing a link to the source material.

Notice that there is a link to the source in the caption below the image.

How to Complete this Quiz
Download the file called figures.zip. You'll find index.html and an image inside.
Edit index.html so that the caption displays below the image. I want you to use a relative path to the image, which is going to be in the same directory as index.html. Remember, a relative path points to the file from the current directory - it does not start with the root. (Your relative path should not start with /, file://, C:\, or http://).
There's more than one option for completing this quiz but you only have to complete one option! You'll get a different code depending on how you complete it. Type the code into the box in the next screen to pass the quiz!