Add files via upload
Mockup to Website
It's common for web developers to work with designers who focus on creating user interfaces and user experiences. Designers use software like Adobe Photoshop to mock up - draw - websites. The mockups that they create are usually just images of websites with some annotations and descriptions.

As a web developer, one of the tasks you might be asked to do is take a mockup created by a visual designer and translate it into a live website. I use the word "translate" because the process of going from mockup to website is similar to the process of translating between natural languages. Just as you can create the same meaning using different words and phrases with a natural language, you can create the same website design using different HTML elements.

I want you to practice the process of going from a mockup to a website now! Here is a website mockup (note: I zoomed in for the screen shot):


You can find a copy of this image in web-dev-blog.zip in the instructor notes. You'll also find a file to start editing (index.html) and a copy of the mockup.

There are many ways to turn this mockup into website. As such, you won't be getting feedback on your site using the Chrome extension. Rather, I want you to compare your website to the mockup. You'll know you've finished this exercise when your site looks the same :)

How to Complete this Quiz
Download and unzip web-dev-blog.zip.
Open the mockup and decide how you want to create the article.
Edit index.html until your website looks identical to the mockup.
Practice indenting children elements. I'll show you how I indented my HTML in the solution.