Add files via upload
Writing Selectors
At this point, you should be familiar with the basic structure of a CSS statement. Every CSS statement is made up of a selector and a declaration block. The selector tells the browser what HTML element we want to style and the declaration block tells the browser what styles need to be applied to that HTML.

css statement

The basic structure of a CSS statement

For this quiz, I want you to focus exclusively on the selector part of a CSS statement. To do this, I've created a webpage that is lacking style. The webpage already has ids and classes added to the HTML, but it's missing the right selectors to add the style.

<div id="menu">
  <h1 class="item">Chicken Clay Pot</h1>
  <img src="img/clay-pot.jpg" alt="clay pot" class="picture">
  <p class="description">Crispy rice baked in clay pot topped with chicken and vegetables</p>
</div>
/* missing id */ {
  text-align: center;
}
/* missing class */ {
  color: red;
}
/* missing class */ {
  border-radius: 5px;
}
/* missing class */ {
  font-style: italic;
}
It's your job to download the webpage and fill-in the missing selectors. If you do it right, your webpage should end up looking like this...

writing selectors solution

How to Complete this Quiz
Download the zip file called writing-selectors.zip from the resources section or from here.
Open index.html and turn on the Udacity Feedback Chrome Extension.
Edit index.html and add the missing id and class selectors within the <style> tags (alternatively, you can use Developer Tools to add the missing id and class selectors to pass all the tests).
When you've successfully added the correct selectors, copy and paste the code that appears into the next page to finish!
When you're ready to get started, click "Skip to Quiz".

Start Quiz
How to Complete this Quiz
Download the zip file called writing-selectors.zip in the downloadables section.
Open index.html and turn on the Udacity Feedback Chrome Extension.
Edit index.html and add the missing id and class selectors within the <style> tags (alternatively, you can use Developer Tools to add the missing id and class selectors to pass all the tests).
When you've successfully added the correct selectors, copy and paste the code that appears into the next page to finish!