Add files via upload
Link to a Stylesheet
For your last quiz, you'll be linking a stylesheet to a familiar webpage from this lesson.

link to a stylesheet quiz

Do you remember this webpage from the very first video in this lesson?

I've already provided you with the HTML and CSS in the downloadables section. It's your job to download the files and connect the stylesheet to the webpage.

<html>
<head>
  <title>Link to a Stylesheet Quiz</title>
  <!-- add link here -->
</head>
<body>
  <!-- ... -->
</body>
</html>
In the download, styles.css (the stylesheet) is conveinetly tucked away in a folder called css. So, you will need to draw upon your knowledge of paths in order to correctly link the stylesheet to the HTML. If you've done everything correctly, your link should be formatted like...

<link href="path-to-stylesheet/stylesheet.css" rel="stylesheet">
How to Complete this Quiz
Download the zip file called link-to-a-stylesheet.zip from the Supporting Materials below or from here.
Edit the index.html file by adding a link to the missing stylesheet.
View index.html in the browser to see if the stylesheet was applied.
Turn on the Udacity Feedback Chrome Extension (if you haven't already).
When you've successfully added the stylesheet, a code will appear in the extension. Paste the code into the classroom in the next screen.